+++ src/components/CategoryContent.module.scss
