+++ src/components/CategoryContentPanel.module.scss
