--- src/components/CodeList.module.scss
@@ -1,10 +1,0 @@
-@use "style/mixin";
-@use "style/function";
-
-.wrapper {
-  margin-bottom: function.em(16, -24);
-}
-
-.item {
-  margin-bottom: function.em(16, 24);
-}
