+++ src/models/CategoryContent.ts
@@ -1,0 +1,7 @@
+import Model from "./Model";
+import { PageCategory } from "./PageCategory";
+
+export interface CategoryContent extends Model {
+  title: string;
+  Categories: PageCategory[];
+}
