+++ src/models/News.ts
@@ -1,16 +1,16 @@
 import Model from "./Model";
 
-interface News extends Model {
-  date: string;
-  newsContent: NewsContent[];
-}
-
 interface NewsContent {
   fieldId: string;
   title: string;
   details: string;
   linkName: string;
   url: string;
+}
+
+interface News extends Model {
+  date: string;
+  newsContent: NewsContent[];
 }
 
 export default News;
