+++ src/models/PageCategory.ts
@@ -1,10 +1,4 @@
-import Model from "./Model";
 import Post from "./Post";
-
-export interface CategoryContent extends Model {
-  title: string;
-  Categories: PageCategory[];
-}
 
 export interface PageCategory {
   fieldId: string;
